demo/requirements.txt
streamlit>=1.28.0
gradio>=4.0.0
pillow>=10.0.0

